Remove troubleshooting sections
This information was added for the first version when k8s was only with ESP only and not Viya. Instructions on how to refresh and remove ESP Servers are in the main guides. We also worked on making error reporting better so users don't need to be k8s admin to figure out what is going on. 

I also fixed broken links but I could only get the filebrowser to go to master root page. Katja?